feat(capacity): online enospc children if there is enough capacity in the pool
It's not trivial to find out if we should online a child as the pool allocation can be forever
changing.
For capacity v1 start with a kiss approach of simply onlining children if we have *some* space.
We take into account the current replica and volume allocation.
Example, if the failed child's replica has been allocated up to 90% of the volume full capacity,
then we don't need free space for the whole volume capacity.
At the same time, if the volume current allocation is 90% then we need to make sure we have free
space to reach 90% of the volume allocation.
A dummy slack is added on top, which is mostly a placeholder for a future replacement that would
take into account pool allocation growth.

Signed-off-by: Tiago Castro <tiagolobocastro@gmail.com>