fix(capacity/thin): don't fault largest replica if we can't move it
Changes the faulting logic of the largest enospc replica. The replica is now moved first and only
then it is faulted&deleted.
Otherwise we might end up faulting all enospc replicas as each second largest replica eventually
becomes the largest replica :)

We might still need a way to delete the largest replica even if we can't move it as it might allow
us to get enough space for the other replicas.

Signed-off-by: Tiago Castro <tiagolobocastro@gmail.com>