fix(switchover): adds several corner case fixes
Replace path is not idempotent in case the connection was established without moving path.
gRPC Timeouts for connection are not enough which leads into moving the target
when not necessary.
Report of failed paths does not differentiate errors at all.
Report of failed paths returns only 1 error.. we should have an error per path as we report.
 in batches.
In case of no resources for republish volume the ha cluster retries in a loop with no delay.
or backoff..
In error case we don’t clean up shutdown targets.
On switchover failure, failed paths are not cleared in the cluster agent.

Signed-off-by: Tiago Castro <tiagolobocastro@gmail.com>