feat(snapshots/pstor): add initial snapshot definitions
These are the spec-like definitions that are stored in the pstor (etcd).
As a reminder, we use a log-based approach where we first write-out what the operation is,
eg: create, and upon success we complete by writing the final object. This allows us to
recover in case of a crash for example.

It may be that we have only Create&Delete operations for a snapshot, it's not clear to me yet
if we need an Update operation when using clones, but that is TBD.

Signed-off-by: Tiago Castro <tiagolobocastro@gmail.com>