feat(snapshot/create): add core agent create volume snapshot logic
Implements the core logic for creating a volume snapshot.
A snapshot resource is created for creation or retry and a new txn id is allocated
for this operation.
todo: should we start destroying previous txn id here?

If the volume is published then creation goes via the nexus rather than the replica, but
in either case we prepare a ReplicaSnapshot so we can keep track of the operation. This
allows us, for example, to clean up later if we fail/crash/etc.

todo: on retries, check if the txn_id has already been created! (in case the ctrl-plane crashes).

Signed-off-by: Tiago Castro <tiagolobocastro@gmail.com>