feat(snapshot/state): retrieve the replica snapshot states
Add pool id to the replica snapshot spec. Why? Consider what happens when the source volume is
deleted, we no longer have the replica so we cannot determine where the snapshot resides?!
Using the pool id allows to determine where the snapshot resides.
todo: Should we use the pool uuid instead?
todo: refactor VolumeSnapshotState to be able to represent a snapshot state

Signed-off-by: Tiago Castro <tiagolobocastro@gmail.com>