fix(snapshot/retry): add txnid to the snapshot name
This allows for safer retries as the snapshot name will not clash.
Add timeout test which leads to a snapshot in the dataplane, but not in the control-plane.
The retry on the control-plane should then succeed creating a new snapshot in the data-plane.
todo: clean-up previous snapshot transaction in creating status?

Signed-off-by: Tiago Castro <tiagolobocastro@gmail.com>