feat(snapshot/clone): create volume from snapshot clone
Refactor the volume creation logic to use a new trait which allows code reuse between
creating new volumes from "pool" and from a snapshot.
Add logic to select a replica snapshot and create a clone from it.
The name of the clone is the same as its uuid.
todo: start naming replicas/clones with a useful name, maybe part volume id, part replica?

Signed-off-by: Tiago Castro <tiagolobocastro@gmail.com>