fix(snapshot/timeout): increase nexus and replica snapshot timeout
Increase timeouts for nexus and replica snapshots as it seems sometimes the
subsystem pause can take a few seconds. On a particular test it took 11s
to pause.
If we increase the IO timeout this might take even longer? So we probably
need a way to undo snapshot creation because when we timeout the snapshot
might not be created yet so the control-plane can't undo it!

Signed-off-by: Tiago Castro <tiagolobocastro@gmail.com>