fix(csi-node/nvmf/fixup): fixup correct nvme controller
When we replace an existing path, the new path has a different controller number. And so
the controller number and device number now mismatch, meaning we can not safely deref
/sys/class/nvme/nvme{major}
Instead, we can simply deref
/sys/class/block/nvme{major}c*n1/queue
The major ensures we use the original device number, and the glob ensures we modify the
timeout for all controllers.

Signed-off-by: Tiago Castro <tiagolobocastro@gmail.com>