fix(nexus/add-child/v1): make add child v1 idempotent
When v1 nexus add child was added, it was not made idempotent.
Even though this is not an issue per se, as the child eventually gets
GCd and re-added it can cause strange logging..
TODO: should we have different behaviour depending on the state?
Example if faulted should we remove/readd?
Bonus: Fixes old test which stopped working a long time ago when
pstor was enabled for the data-plane by not enabling it for that
particular test only..

Signed-off-by: Tiago Castro <tiagolobocastro@gmail.com>