feat(volume/resize): don't allow volume resize in presence of snapshots
If a volume is expanded after taking a snapshot, then the IOs to the
new io-units(block ranges) fails because the lba range falls out of
bound for replica's parent device(snapshot). Till the issue is fixed,
this check ensures we don't try to expand a volume if there is snapshot
present. SPDK issue #3108.

Signed-off-by: Diwakar Sharma <diwakar.sharma@datacore.com>