fix(agents/ha/cluster): increase path replacement timeout
In a system test run we've seen the disconnect of an old path take 50s
to complete.
We're not sure why this is the case as we cannot reproduce it yet but
for now we can increase the timeout.
This is probably fine as if we can't talk to the node where the app is
then we can't do much more anyway.

We should also consider wether we need to wait for the disconnect at
all. If we can find out if the disconnect has been started, then we
can probably just complete that async and succeed the replace!

Signed-off-by: Tiago Castro <tiagolobocastro@gmail.com>