fix(agents/ha/node): complete reconnect when path is deleting
During the reconnect of a new path, the old path might take some time to
disconnect.
Rather than wait for this (and timeout, which leads to retries) simply
complete call as soon as path is deleting.

In case of deadline exceeded attempt replace a few times before trying
to republish with a new target.

Signed-off-by: Tiago Castro <tiagolobocastro@gmail.com>