build(openapi): don't depend on version.txt
Not quite sure why the version.txt was added since we know that we should not
be adding output files as an input to the build.rs of a crate.
This is causing extra rebuilds during CI!
I suspect it may have been added to handle git clean of the openapi dir!
I think the correct way to handle that is to simply move the output of the
build.rs to the target folder where it should have lived :-)
Let's remove this and then pay attention to missing rebuilds..

Signed-off-by: Tiago Castro <tiagolobocastro@gmail.com>