test(bdd): ensure nvme disconnect in failure
When we connect to a device but we fail to wait for it to come up, we end up
not running the teardown logic (in the yield), as that code is only run in
case of no error.
Instead, let's catch any error, disconnect and re-reaise the error.
Also add some logging to help catch what the error actually is..

Signed-off-by: Tiago Castro <tiagolobocastro@gmail.com>