feat(deployer): add idle io-engines
These are useful to sort of simulate upgrades.
This is required because there's no easy way of updating a container image tag.
A better alternative to add in the future would be to extend composer to add
containers to existing cluster and extending cluster to allow for this via its
components as well.

Signed-off-by: Tiago Castro <tiagolobocastro@gmail.com>