fix: don't add replica to nexus which can't rebuild properly
Check if all nodes have the rebuild fix for the following operations:
- increase replica count on volume
- move replica
- add replica to nexus
- take snapshot on nr volume

Checking all nodes might lead to some false positives and also some issues if some nodes
have been removed from power but not from the pstor, but it's a very
simple way of ensuring all nodes are fixed.
A potential improvement may be to check if the replica nodes have the fix.

Signed-off-by: Tiago Castro <tiagolobocastro@gmail.com>