fix: don't disown replica from unshutdown nexus
In case nexus shutdown failed we used to disown the replica from
both the volume and the nexus.
This can be a problem if the nexus is still running as the io-engine
does not handle it gracefully, leading into pool lock issues.
Instead, simply disown the replica from the volume and not the nexus.

Signed-off-by: Tiago Castro <tiagolobocastro@gmail.com>