feat: handle data integrity in republish path
When a new nexus is created in republish path, it can pick up a healthy
replica that is local to nexus node. But it is possible that the old
nexus could momentarily come back and fails an IO to this replica and in
turn marking it unhealthy in etcd. Since new nexus has already chosen
this replica, it goes ahead using this as source of truth and hence
compromising data integrity. This change lets core-agent indicate to old
nexus that it should self shutdown now without modifying updating
anything in etcd.

Signed-off-by: Diwakar Sharma <diwakar.sharma@datacore.com>