fix: remove volume health entries on delete
When a volume is deleted, destroy all related health entries.
It has been observed that upon republished, the previous info
is not deleted, so even if we fix that a lot of entries may
already be leaked, so let's ensure we delete all entries for
a given volume on volume deletion.

Signed-off-by: Tiago Castro <tiagolobocastro@gmail.com>