feat(rest-plugin): delete volume by uuid
Add delete volume subcommand to the rest-plugin utility.
A prompt is used to ensure the user really intends to delete
the volume and there's no way back, cannot be undone!
The K8s plugin utility will add additional checks, such as
ensuring the equivalent PV is already deleted.

In the future we may be able to improve this a bit by adding
another staging area or garbage bin, where deleted resources
could be kept for up to X days.

Signed-off-by: Tiago Castro <tiagolobocastro@gmail.com>