refactor(pstor/etcd-watcher): remove watch status update
The keep alive is already useful for triggering watch restarts in case the
connection is in a zombie state.
Also, to address what happens when no progress status is received within
the timeout, we can also issue an explicit request for progress and update
the internal rev this way. This is also a much safer way of updating the
rev as there might some race between getting the status and watches?

Signed-off-by: Tiago Castro <tiagolobocastro@gmail.com>