fix(ha/multipath): assume multipath as ON
If the path /sys/module/nvme_core/parameters/multipath is not found
but /sys/module/nvme_core/parameters/iopolicy is then assume multipath
is enabled.

A recent change from RHEL10 made multipath enabled by default, but also
removed the multipath parameter for some reason, which is a breaking
change for us!

On very old kernels it's perhaps possible that multipath parameter is
not found but nvme_core is enabled and so we check the iopolicy which
in the systems I checked is present.

Signed-off-by: Tiago Castro <tiagolobocastro@gmail.com>