feat(simulation): add simulator of cluster bundles
The kubectl-mayastor plugin takes a support bundle which includes various
resources, such as pools and volumes.
Sometimes it's helpful to not only visualize but also run commands against
a similar configured cluster.
In order to aid this, we add a simulation tool here, which is able to load
a bundle and attempts to re-create similar spec and state.
The process is not entirely accurate because the bundle is taken using the
public api, and here we are using the internal types.

Signed-off-by: Tiago Castro <tiagolobocastro@gmail.com>