refactor(pool/scheduling): use more precise snapshot errors
When a pool is cordoned the returned error should be failed precondition.
When pools are filtered out, we end up with no choices, and return a
hardcoded generic error of resource exhausted.

In order to achieve this, we have to remember what reasons were used to
exclude a pool.
This changes introduces a list of excluded resources when they are filtered
out.
For the moment, this has just been done for the pools, though it could be
extended to other schedulers.

Signed-off-by: Tiago Castro <tiagolobocastro@gmail.com>