fix(plugin): sort tabled output on first column
Add sort_rows method to CreateRows and CreateRow traits.
The first column is typically the identifier of something so default
to sorting by row. This may be opted out ofcourse.

Certain outputs such as the topologies are a bit complicated so have
to be sorted manually prior.

This approach has the downside that the json/yaml output is not sorted,
but that's probably fine for now.

Signed-off-by: Tiago Castro <tiagolobocastro@gmail.com>