fix(node-spread): use spread policy on replica add/remove
Use node-spread policy when adding new replicas to a volume.
This is done by updating the existing filtering logic, ensuring that
new replicas are created on nodes which contain the spread keys and
also ensuring that the new values are different from existing replicas.

Also when removing replicas we now use the spread, preferring to remove
replicas which don't contain the spreads keys (all else being equal).
If more than 1 replicas don't contain the spread keys, the deadlock is
broken by picking the one with more spread value collisions.

Signed-off-by: Tiago Castro <tiagolobocastro@gmail.com>