feat(affinity-group): support scale down to 1 replica
Scaling down affinity groups to 1 replica was previously blocked because
we didn't want to get into a situation where replicas would end up
sharing the same node, defeating the purpose of Affinity Groups!

We now loosen this restriction by allowing the scale down if we don't
get into that scenario :)

This is done by creating a list of restricted nodes which are nodes where
a 1-replica volume part of the same affinity group places its replica.
In such case, we don't want to scale down in such a way that the last
remaining replica would be placed in restricted nodes!

This isn't the full monty as replicas may be sharing nodes when Scaling
back from 3 to 2, which would end up impeding further scaling down to 1!
For this we'd have to implement some logic to improve the sorting.
This limitation is fine for now, we can document scaling up and down as
a way of moving replicas around!

Signed-off-by: Tiago Castro <tiagolobocastro@gmail.com>
(cherry picked from commit a4d63e705d50021a28be8ca1f3e3461dea8d21d6)